Update autorip.service file for local env
This commit updates the `autorip.service` file to reflect the location
of local volume storage.  Docker image name is updated as well.

Please note that the built docker image is not currently published to
dockerhub.  So, the image will need to be built locally using the
following command prior to inserting a disc to kick off this Exec.

```
sudo docker build -t johnduarte/abcde-autorip .
```