Use master tag of phusion/baseimage
Update to use master tag of phusion/baseimage in order to pick up
abcde version 2.9+ which includes support for musicbrainz api v2.
Musicbrainz is deprecating api v1 which causes abcde to never find
metadata if using version 2.8.x or below.

The master tag of phusion/baseimage is using Ubuntu focal which
includes abcde version 2.9.3.  The production release of this docker
image has not been released yet.  After that happens, the image in
the Dockerfile should be updated to reflect the production tag.